connector: code separation from __init__
* Moves package `__init__.py` to `connector.py` to pave the way for new
  REST API client code.  (closes #6)
